+++ ApplyToBecomeInternal/ApplyToBecome.Data/Models/Establishment/AddressResponse.cs
@@ -1,0 +1,12 @@
+﻿namespace ApplyToBecome.Data.Models.Establishment
+{
+	public class AddressResponse
+	{
+		public string Street { get; set; }
+		public string Locality { get; set; }
+		public string AdditionalLine { get; set; }
+		public string Town { get; set; }
+		public string County { get; set; }
+		public string Postcode { get; set; }
+	}
+}
