+++ ApplyToBecomeInternal/ApplyToBecome.Data/Models/Establishment/CensusResponse.cs
@@ -1,0 +1,8 @@
+﻿namespace ApplyToBecome.Data.Models.Establishment
+{
+	public class CensusResponse
+	{
+		public string NumberOfPupils { get; set; }
+		public string PercentageFsm { get; set; }
+	}
+}
