+++ ApplyToBecomeInternal/ApplyToBecome.Data/Models/Establishment/EstablishmentResponse.cs
@@ -1,0 +1,24 @@
+﻿namespace ApplyToBecome.Data.Models.Establishment
+{
+	/// <remarks>
+	/// Copy of TramsDataApi.ResponseModels.EstablishmentResponse with most properties omitted
+	/// </remarks>
+	public class EstablishmentResponse
+	{
+		public string Urn { get; set; }
+		public string LocalAuthorityCode { get; set; }
+		public string LocalAuthorityName { get; set; }
+		public string EstablishmentNumber { get; set; }
+		public string EstablishmentName { get; set; }
+		public NameAndCodeResponse EstablishmentType { get; set; }
+		public NameAndCodeResponse PhaseOfEducation { get; set; }
+		public string StatutoryLowAge { get; set; }
+		public string StatutoryHighAge { get; set; }
+		public NameAndCodeResponse Diocese { get; set; }
+		public string SchoolCapacity { get; set; }
+		public CensusResponse Census { get; set; }
+		public string OfstedLastInspection { get; set; }
+		public NameAndCodeResponse ParliamentaryConstituency { get; set; }
+		public MISEstablishmentResponse MISEstablishment { get; set; }
+	}
+}
