+++ ApplyToBecomeInternal/ApplyToBecome.Data/Models/Establishment/MISEstablishmentResponse.cs
@@ -1,0 +1,13 @@
+﻿namespace ApplyToBecome.Data.Models.Establishment
+{
+	public class MISEstablishmentResponse
+	{
+		public string OverallEffectiveness { get; set; }
+		public string QualityOfEducation { get; set; }
+		public string BehaviourAndAttitudes { get; set; }
+		public string PersonalDevelopment { get; set; }
+		public string EffectivenessOfLeadershipAndManagement { get; set; }
+		public string EarlyYearsProvision { get; set; }
+		public string SixthFormProvision { get; set; }
+	}
+}
