+++ ApplyToBecomeInternal/ApplyToBecome.Data/Models/Establishment/NameAndCodeResponse.cs
@@ -1,0 +1,8 @@
+﻿namespace ApplyToBecome.Data.Models.Establishment
+{
+	public class NameAndCodeResponse
+	{
+		public string Name { get; set; }
+		public string Code { get; set; }
+	}
+}
